requirements.txt
Django==2.1
pytz==2018.5
djangorestframework==3.8.2